requirements.txt
Django==2.1.3
google-cloud-language
google-cloud-speech
djangorestframework
